fix(toolkit-lib): `CloudAssemblySourceBuilder` and `BaseDeployOptions` are not part of the public API
