+++ packages/@aws-cdk/toolkit-lib/lib/api/cloud-assembly/private/index.ts
@@ -2,5 +2,4 @@ export * from './context-aware-source';
 export * from './stack-assembly';
 export * from './exec';
 export * from './prepare-source';
-export * from './source-builder';
 export * from './stack-selectors';
