+++ packages/@aws-cdk/toolkit-lib/lib/toolkit/toolkit.ts
@@ -20,7 +20,7 @@ import { BootstrapSource } from '../actions/bootstrap';
 import { AssetBuildTime, type DeployOptions } from '../actions/deploy';
 import {
   buildParameterMap,
-  type ExtendedDeployOptions,
+  type PrivateDeployOptions,
   removePublishedAssetsFromWorkGraph,
 } from '../actions/deploy/private';
 import { type DestroyOptions } from '../actions/destroy';
@@ -40,7 +40,8 @@ import { Bootstrapper } from '../api/bootstrap';
 import type { ICloudAssemblySource } from '../api/cloud-assembly';
 import { CachedCloudAssembly, StackSelectionStrategy } from '../api/cloud-assembly';
 import type { StackAssembly } from '../api/cloud-assembly/private';
-import { ALL_STACKS, CloudAssemblySourceBuilder } from '../api/cloud-assembly/private';
+import { ALL_STACKS } from '../api/cloud-assembly/private';
+import { CloudAssemblySourceBuilder } from '../api/cloud-assembly/source-builder';
 import type { StackCollection } from '../api/cloud-assembly/stack-collection';
 import { Deployments } from '../api/deployments';
 import { DiffFormatter } from '../api/diff';
@@ -409,7 +410,7 @@ export class Toolkit extends CloudAssemblySourceBuilder {
   /**
    * Helper to allow deploy being called as part of the watch action.
    */
-  private async _deploy(assembly: StackAssembly, action: 'deploy' | 'watch', options: ExtendedDeployOptions = {}): Promise<DeployResult> {
+  private async _deploy(assembly: StackAssembly, action: 'deploy' | 'watch', options: PrivateDeployOptions = {}): Promise<DeployResult> {
     const ioHelper = asIoHelper(this.ioHost, action);
     const selectStacks = options.stacks ?? ALL_STACKS;
     const synthSpan = await ioHelper.span(SPAN.SYNTH_ASSEMBLY).begin({ stacks: selectStacks });
@@ -1139,7 +1140,7 @@ export class Toolkit extends CloudAssemblySourceBuilder {
   ): Promise<void> {
     // watch defaults to hotswap deployment
     const deploymentMethod = options.deploymentMethod ?? { method: 'hotswap' };
-    const deployOptions: ExtendedDeployOptions = {
+    const deployOptions: PrivateDeployOptions = {
       ...options,
       cloudWatchLogMonitor,
       deploymentMethod,
